+++ shell.nix
@@ -4,5 +4,6 @@ pkgs.mkShell {
     ocaml 
     ocaml-lsp
     ocamlformat_0_26_2
+    utop
     dune_3 findlib camlp-streams alcotest qcheck qcheck-alcotest; };
 }